requirements.txt
python-dotenv
python-telegram-bot
schedule
pillow
groq
stability-sdk
feedparser
